Add interface support
